Interview Add Question

+++ src/ExampleCode/Test.java
@@ -2,10 +2,10 @@ package ExampleCode;
 
 public class Test {
     public static void main(String[] args) {
-        Double a = 1.0;
-        Double b = 1.0;
-        Double c = 99999.0;
-        Double d = 99999.0;
+        Integer a = 1;
+        Integer b = 1;
+        Integer c = 99999;
+        Integer d = 99999;
 
         System.out.println(a == b);
         System.out.println(c == d);
